Move filestate locking to apply time
This aligns the filestate behaviour with httpstate in that we won't take
a lock during the preview phase of an update. Instead we take the lock
at the point we start doing a non-preview apply

We've tried this once before
(https://github.com/pulumi/pulumi/pull/8806) but got a couple of user
reports about locking not working soon after it released, and so
reverted it just in case. Further investigation into those user issues
suggested they we're due to user fault, not this change.

There's one comment from myself on Slack from around that time saying
I'm not sure this change is correct, but even with that nudge and
looking at it with fresh eyes I think it is. But worth having a hard
think about this before hitting approve.