Send resource name and type as well as URN to providers
We want to make URN's opaque tokens that don't need clients to parse them. One
thing providers currently parse the URN for is the resource name and type.
Pretty simple to extend the rpc interface to send these tokens as individual
fields plus the URN.

We can start getting providers to use these tokens and fallback to parsing,
eventuall providers can decide to take a requirement on an up-to-date engine
that sends these fields.