Send args from convert to converter plugin
This matches up with how we send args to converter plugins from `import`. Looks
like this will be useful in at least two places already. Firstly the bicep
converter needs to know which file to convert. Secondly the terraform convert
needs a --convert-examples (or the like) flag to turn on it's bulk converter
mode for the terraform bridge to use.

Also fixes that we we're limiting `import` to three args, but a converter may take more.