More efficent mapping lookup
Inspired by a comment from Zaid. This allows providers to return what providers
they have mapping information for without having to marshal all their mapping
data to the engine at the same time, this could save transmitting a lot of data
that the engine might not ever need (for example if it's not converting code
for that specific provider).  It also allows provider to support mulitple
mappings.