Move SecretsManager from SnapshotPersister to SnapshotManager
It should never have really been the role of the persister to decide what
SecretsManager to use, and they were in practice always being setup with a
SnapshotManager anyway so not a lot of code to move the secrets manager around.
This also gets the httpstate and filestate backends in alignment over which
SecretsManager variable they're going to use.