[sdkgen/python] Fix error calling `_configure`
In order to set default values on passed-in dictionaries for object args, we check to see if the value isn't an instance of the args class, and then proceed with treating it as a dictionary. However, if the value is optional, it will be `None` and therefore not an instance of the args class, so we proceed with calling configure on it.

If the object itself has any required args, this will then fail, because we'd be passing an empty dict to `_configure`.

The fix is to check both the value is not None and it is not an instance of the args class, before treating it as a dict.