[sdk/go] Ensure Assets of AssetArchive are non-nil when creating and deserializing (#14007)
Fixes #11103

The problem in the issue the distinction between `Assets` or an archive
being `nil` vs. being an empty map of assets. When creating and
deserializing Archives, `nil` was allowed which leads to panics when
asserting `archive.IsAssets()` because the `Assets` is `nil` even though
it is a proper archive.

This PR fixes the issue by ensuring that an instance of `Archive` will
have a non-nil `Assets` field whether that is during construction or
marshalling.

## Checklist

- [ ] I have run `make tidy` to update any new dependencies
- [x] I have run `make lint` to verify my code passes the lint check
  - [x] I have formatted my code using `gofumpt`

<!--- Please provide details if the checkbox below is to be left
unchecked. -->
- [x] I have added tests that prove my fix is effective or that my
feature works
<!--- 
User-facing changes require a CHANGELOG entry.
-->
- [x] I have run `make changelog` and committed the
`changelog/pending/<file>` documenting my change
<!--
If the change(s) in this PR is a modification of an existing call to the
Pulumi Cloud,
then the service should honor older versions of the CLI where this
change would not exist.
You must then bump the API version in
/pkg/backend/httpstate/client/api.go, as well as add
it to the service.
-->
- [ ] Yes, there are changes in this PR that warrants bumping the Pulumi
Cloud API version
<!-- @Pulumi employees: If yes, you must submit corresponding changes in
the service repo. -->