Stop defensive copying of entire event payloads
The deep copy of engine payloads was originally added because certain operations during display mutate engine events. However, this has become problematic as internals of the payload itself can mutate concurrently with operations in the engine, leading to panics due to concurrent iteration and writing of maps.

Rather than take a defensive copy of the entire payload, only make a defensive copy of the specific things being mutated inside display.