Add ability for transforms to convey diagnostics
In the case where a transform runs against a resource with
unknown properties, we want to be able to issue a warning, but
still proceed, to match behavior we have for policies.

Since transforms don't convey diagnostics in the same manner as
policies, intentionally, this change lets them return it as an
independent diagnostic message per transform result.