Fix Homebrew install test (#14082)
Our cron job that tests installing Pulumi has been failing when
installing from Homebrew on macOS. The `pulumi` package depends on `go`,
and Homebrew fails installing `go` due to conflicting files on the
GitHub action runner. This happens because the runner image already has
Go installed in the same places that Homebrew is trying to install go.

The workaround is to delete the conflicting Go files on the runner
before running the Homebrew install.

A successful test run with this fix is available here:
https://github.com/pulumi/pulumi/actions/runs/6383582011/job/17324492547

Fixes #13938