schema, eval: support anyOf, oneOf, $defs, $ref (#37)
Add support to the schema package and the validator for anyOf, oneOf,
$defs, and $ref.

When dealing with unknown values, the validator treats oneOf as if it
were anyOf (i.e. it requires only that _some_ element of the oneOf
validates the input, rather than _exactly one_ element). This is
consistent with the generally liberal approach to validation taken by
the evaluator when dealing with unknowns.

When dealing with concrete values, the usual JSON schema semantics
apply.

These changes also add some additional tests to bring some additional
coverage to the evaluator.