fix MockMonitor reporting DeletedWith wasn't supported
We should support all features in mocks that we are supporting
otherwise.  We did a similar thing in pulumi-dotnet:
https://github.com/pulumi/pulumi-dotnet/pull/93.

However here we're not simply changing hasSupport to true.
https://github.com/pulumi/pulumi/commit/1e09626bc731ad2379d646b0b3b46036e8a7661e
has a good explanation of why we only support specific values here.
However really we only need to disable support for `outputValues`, and
that only if output values are disabled.  Originally this was true
if `PULUMI_ENABLE_OUTPUT_VALUES` was set, but that environment
variable was later inverted: 3027d01f259d304abb764a0e8793ed2996cede56.