[cli/engine] Restore elided asset contents from returned inputs from Read operations (#14078)
# Description

When we send inputs to `Read` operations due to `pulumi refresh` we skip
sending asset contents. However, `Read` implementation might return the
asset-valued input properties unchanged (without their asset contents)
and the engine will simply write out the empty assets to the state
causing an invalid shape of asset in the inputs of resources.

This PR implements a function `restoreElidedAssetContents` and uses it
in the `Read` implementation on the new inputs and new state returned
from the provider plugins.

Testing the CLI locally using this fix, it does correctly write out the
asset contents to the state _after_ a `pulumi refresh`. ~However, when
running `preview` after `refresh`, the nodejs SDK still fails with the
same error in #6691 because `RegisterResource` still returns the inputs
that contain empty/invalid assets~

**EDIT**: Fixed the issue with the inputs sent to the nodejs SDK are
missing asset contents (thanks @Frassle 🙏)

Fixes #6691 

### TO-DO
- [x] Figure out why `RegisterResource` is sending empty assets for
during `preview` after `refresh`

## Checklist

- [ ] I have run `make tidy` to update any new dependencies
- [x] I have run `make lint` to verify my code passes the lint check
  - [x] I have formatted my code using `gofumpt`

<!--- Please provide details if the checkbox below is to be left
unchecked. -->
- [x] I have added tests that prove my fix is effective or that my
feature works
<!--- 
User-facing changes require a CHANGELOG entry.
-->
- [x] I have run `make changelog` and committed the
`changelog/pending/<file>` documenting my change
<!--
If the change(s) in this PR is a modification of an existing call to the
Pulumi Cloud,
then the service should honor older versions of the CLI where this
change would not exist.
You must then bump the API version in
/pkg/backend/httpstate/client/api.go, as well as add
it to the service.
-->
- [ ] Yes, there are changes in this PR that warrants bumping the Pulumi
Cloud API version
<!-- @Pulumi employees: If yes, you must submit corresponding changes in
the service repo. -->