Move language check from installPlugin to installPlugins (#14103)
<!--- 
Thanks so much for your contribution! If this is your first time
contributing, please ensure that you have read the
[CONTRIBUTING](https://github.com/pulumi/pulumi/blob/master/CONTRIBUTING.md)
documentation.
-->

# Description

<!--- Please include a summary of the change and which issue is fixed.
Please also include relevant motivation and context. -->

As part of looking into abstracting away the plugin checks for engine
tests, I noticed we have two _very_ similar methods for install plugins.
The main difference was that in `installPlugin` we skipped the install
if the plugin was a language plugin.

This isn't the ideal behaviour, so I've moved that check up into
`installPlugins` and checked for "bundleness" not just being a language
plugin and made it an error (similar error to what people get if they
try and `pulumi plugin install language dotnet`).

So now if the engine wants to use a plugin that should be bundled but
it's missing we'll just error out the deployment.

A follow up PR will look at merging `InstallPlugin` and `installPlugin`
into one method.

## Checklist

- [x] I have run `make tidy` to update any new dependencies
- [x] I have run `make lint` to verify my code passes the lint check
  - [x] I have formatted my code using `gofumpt`

<!--- Please provide details if the checkbox below is to be left
unchecked. -->
- [ ] I have added tests that prove my fix is effective or that my
feature works
<!--- 
User-facing changes require a CHANGELOG entry.
-->
- [x] I have run `make changelog` and committed the
`changelog/pending/<file>` documenting my change
<!--
If the change(s) in this PR is a modification of an existing call to the
Pulumi Cloud,
then the service should honor older versions of the CLI where this
change would not exist.
You must then bump the API version in
/pkg/backend/httpstate/client/api.go, as well as add
it to the service.
-->
- [ ] Yes, there are changes in this PR that warrants bumping the Pulumi
Cloud API version
<!-- @Pulumi employees: If yes, you must submit corresponding changes in
the service repo. -->