Summarize remediations by default
So as not to overwhelm the end user (e.g., in the case that
every resource in a 100 resource stack gets remediated), we summarize
the output by default. Notably, we show output like this:

    aws-auto-tagger@v0.0.1
     - [remediate] s3-tags (7 resources)

If you want to see all seven lines independently, you can pass the
--show-policy-remediations flag.

All resources are still shown if you expand the preview's details.